package is now in a publishable state
