+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "readability",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "description": "A standalone version of the readability library used for Firefox Reader View.",
-  "main": "Readability.js",
+  "main": "index.js",
   "scripts": {
     "test": "mocha test/index.js",
     "generate-testcase": "node test/generate-testcase.js"
