+++ test/index.js
@@ -1,3 +1,5 @@
+"use strict";
+
 var path = require("path");
 var fs = require("fs");
 var prettyPrint = require("html").prettyPrint;
@@ -5,29 +7,7 @@ var chai = require("chai");
 chai.config.includeStack = true;
 var expect = chai.expect;
 
-// We want to load Readability and JSDOMParser, which aren't set up as commonjs libraries,
-// and so we need to do some hocus-pocus with 'vm' to import them on a separate scope
-// (identical) scope context.
-var vm = require("vm");
-var readabilityPath = path.join(__dirname, "..", "Readability.js");
-var jsdomPath = path.join(__dirname, "..", "JSDOMParser.js");
-
-
-var scopeContext = {};
-// We generally expect dump() and console.{whatever} to work, so make these available
-// in the scope we're using:
-scopeContext.dump = console.log
-scopeContext.console = console;
-
-// Actually load files. NB: if either of the files has parse errors,
-// node is dumb and shows you a syntax error *at this callsite* . Don't try to find
-// a syntax error on this line, there isn't one. Go look in the file it's loading instead.
-vm.runInNewContext(fs.readFileSync(jsdomPath), scopeContext, jsdomPath);
-vm.runInNewContext(fs.readFileSync(readabilityPath), scopeContext, readabilityPath);
-
-// Now make references to the globals in our scope so we can use them easily:
-var Readability = scopeContext.Readability;
-var JSDOMParser = scopeContext.JSDOMParser;
+var readability = require('../index.js');
 
 function readFile(path) {
   return fs.readFileSync(path, {encoding: "utf-8"});
@@ -64,8 +44,7 @@ describe("Test page", function() {
       };
 
       beforeEach(function() {
-        doc = new JSDOMParser().parse(source);
-        result = new Readability(uri, doc).parse();
+        result = readability(uri, source);
       });
 
       it("should return a result object", function() {
